feat: new hero section
