+++ src/app/page.tsx
@@ -18,12 +18,10 @@ const fetchImages = async () => {
     let base64Results = await Promise.all(base64Promises);
     return base64Results;
   };
-  (data.mainSection.blurDataUrl = await getBase64(
-    `${srcPrefix}${data.mainSection.image}`
-  )),
-    (data.missionSection.blurDataUrls = (await getBase64Helper(
-      data.missionSection.images
-    )) as unknown as string[]);
+
+  data.missionSection.blurDataUrls = (await getBase64Helper(
+    data.missionSection.images
+  )) as unknown as string[];
   data.projectSection.blurDataUrls = (await getBase64Helper(
     data.projectSection.images
   )) as unknown as string[];
@@ -36,7 +34,7 @@ const Page = async () => {
     <>
       <WithSmoothScroll>
         <div className="wrapper">
-          <HeroSection {...data.mainSection} />
+          <HeroSection />
         </div>
         <ProjectSection />
         <MissionSection {...data.missionSection} />
