+++ src/components/index.js
@@ -1,10 +1,9 @@
 export { PrimaryLayout } from './layouts/PrimaryLayout';
 export { ContactSection } from './sections/ContactSection';
 export { MissionSection } from './sections/MissionSection';
-export { HeroSection } from './sections/HeroSection';
 export { WithSmoothScroll } from './layouts/WithSmoothScroll';
 export { TestimonialSection } from './sections/TestimonialSection';
 export { SkillSection } from './sections/SkillSection';
 export { ContactCTASection } from './sections/ContactCTASection';
 export { ProjectSection } from './sections/ProjectSection';
-//export { HeroSection } from './sections/HeroNew';
+export { HeroSection } from './sections/HeroSection';
