+++ src/templates/home.yaml
@@ -1,15 +1,5 @@
 heroSection:
-  imageUrl: /images/skateboard.png
-  blurDataUrl:
-mainSection:
-  subheading: I am Alex
-  headingAccent: Beautiful
-  heading: UI/UX
-  description: >
-    I’ve worked on different open-source projects and have earned contributor roles, I am ready to bring my skills and provide the best quality code and take ownership for what I build. I also enjoy many other hobbies from rock climbing,
-    canoeing and skateboarding.
-  image: /images/skateboard.webp
-  blurDataUrl:
+  imageUrl: /images/skateboard-gray.webp
 missionSection:
   heading: Mission
   images:
